procfs_core

Module keyring

Functions related to the in-kernel key management and retention facility

For more details on this facility, see the keyrings(7) man page.

Structs

Key

A key

KeyFlags

Various key flags

KeyUser

Information about a user with at least one key

KeyUsers

Information about a set of users with at least one key.

Keys

A set of keys.

PermissionFlags

Bitflags that represent the permissions for a key

Permissions

Enums

KeyTimeout

KeyType