procfs_core

Module net

Information about the networking layer.

This module corresponds to the /proc/net directory and contains various information about the networking layer.

Structs

ARPEntry

An entry in the ARP table

ARPFlags

Flags for ARP entries

ARPHardware

Hardware type for an ARP table entry.

ArpEntries

ARP table entries.

DeviceStatus

General statistics for a network interface/device

InterfaceDeviceStatus

Device status information for all network interfaces.

RouteEntries

A set of ipv4 routes.

RouteEntry

An entry in the ipv4 route table

Snmp

This struct holds the data needed for the IP, ICMP, TCP, and UDP management information bases for an SNMP agent.

Snmp6

This struct holds the data needed for the IP, ICMP, TCP, and UDP management information bases for an SNMP agent.

TcpNetEntries

TCP socket entries.

TcpNetEntry

An entry in the TCP socket table

UdpNetEntries

UDP socket entries.

UdpNetEntry

An entry in the UDP socket table

UnixNetEntries

Unix socket entries.

UnixNetEntry

An entry in the Unix socket table

Enums

IpForwarding

The indication of whether this entity is acting as an IP gateway in respect to the forwarding of datagrams received by, but not addressed to, this entity. IP gateways forward datagrams. IP hosts do not (except those source-routed via the host).

TcpRtoAlgorithm

The algorithm used to determine the timeout value used for retransmitting unacknowledged octets.

TcpState

UdpState

UnixState