Struct procfs_core::net::Snmp

pub struct Snmp {

    pub ip_forwarding: IpForwarding,
    pub ip_default_ttl: u32,
    pub ip_in_receives: u64,
    pub ip_in_hdr_errors: u64,
    pub ip_in_addr_errors: u64,
    pub ip_forw_datagrams: u64,
    pub ip_in_unknown_protos: u64,
    pub ip_in_discards: u64,
    pub ip_in_delivers: u64,
    pub ip_out_requests: u64,
    pub ip_out_discards: u64,
    pub ip_out_no_routes: u64,
    pub ip_reasm_timeout: u64,
    pub ip_reasm_reqds: u64,
    pub ip_reasm_oks: u64,
    pub ip_reasm_fails: u64,
    pub ip_frag_oks: u64,
    pub ip_frag_fails: u64,
    pub ip_frag_creates: u64,
    pub icmp_in_msgs: u64,
    pub icmp_in_errors: u64,
    pub icmp_in_csum_errors: u64,
    pub icmp_in_dest_unreachs: u64,
    pub icmp_in_time_excds: u64,
    pub icmp_in_parm_probs: u64,
    pub icmp_in_src_quenchs: u64,
    pub icmp_in_redirects: u64,
    pub icmp_in_echos: u64,
    pub icmp_in_echo_reps: u64,
    pub icmp_in_timestamps: u64,
    pub icmp_in_timestamp_reps: u64,
    pub icmp_in_addr_masks: u64,
    pub icmp_in_addr_mask_reps: u64,
    pub icmp_out_msgs: u64,
    pub icmp_out_errors: u64,
    pub icmp_out_dest_unreachs: u64,
    pub icmp_out_time_excds: u64,
    pub icmp_out_parm_probs: u64,
    pub icmp_out_src_quenchs: u64,
    pub icmp_out_redirects: u64,
    pub icmp_out_echos: u64,
    pub icmp_out_echo_reps: u64,
    pub icmp_out_timestamps: u64,
    pub icmp_out_timestamp_reps: u64,
    pub icmp_out_addr_masks: u64,
    pub icmp_out_addr_mask_reps: u64,
    pub tcp_rto_algorithm: TcpRtoAlgorithm,
    pub tcp_rto_min: u64,
    pub tcp_rto_max: u64,
    pub tcp_max_conn: i64,
    pub tcp_active_opens: u64,
    pub tcp_passive_opens: u64,
    pub tcp_attempt_fails: u64,
    pub tcp_estab_resets: u64,
    pub tcp_curr_estab: u64,
    pub tcp_in_segs: u64,
    pub tcp_out_segs: u64,
    pub tcp_retrans_segs: u64,
    pub tcp_in_errs: u64,
    pub tcp_out_rsts: u64,
    pub tcp_in_csum_errors: u64,
    pub udp_in_datagrams: u64,
    pub udp_no_ports: u64,
    pub udp_in_errors: u64,
    pub udp_out_datagrams: u64,
    pub udp_rcvbuf_errors: u64,
    pub udp_sndbuf_errors: u64,
    pub udp_in_csum_errors: u64,
    pub udp_ignored_multi: u64,
    pub udp_lite_in_datagrams: u64,
    pub udp_lite_no_ports: u64,
    pub udp_lite_in_errors: u64,
    pub udp_lite_out_datagrams: u64,
    pub udp_lite_rcvbuf_errors: u64,
    pub udp_lite_sndbuf_errors: u64,
    pub udp_lite_in_csum_errors: u64,
    pub udp_lite_ignored_multi: u64,

}

This struct holds the data needed for the IP, ICMP, TCP, and UDP management information bases for an SNMP agent.

For more details, see RFC1213 and SNMP counter

Fields
ip_forwarding: IpForwardingip_default_ttl: u32

The default value inserted into the Time-To-Live field of the IP header of datagrams originated at this entity, whenever a TTL value is not supplied by the transport layer protocol.

ip_in_receives: u64

The total number of input datagrams received from interfaces, including those received in error.

ip_in_hdr_errors: u64

The number of input datagrams discarded due to errors in their IP headers.

ip_in_addr_errors: u64

The number of input datagrams discarded because the IP address in their IP header’s destination field was not a valid address to be received at this entity.

ip_forw_datagrams: u64

The number of input datagrams for which this entity was not their final IP destination, as a result of which an attempt was made to find a route to forward them to that final destination.

ip_in_unknown_protos: u64

The number of locally-addressed datagrams received successfully but discarded because of an unknown or unsupported protocol.

ip_in_discards: u64

The number of input IP datagrams for which no problems were encountered to prevent their continued processing, but which were discarded (e.g., for lack of buffer space).

ip_in_delivers: u64

The total number of input datagrams successfully delivered to IP user-protocols (including ICMP).

Note that this counter does not include any datagrams discarded while awaiting re-assembly.

ip_out_requests: u64

The total number of IP datagrams which local IP user-protocols (including ICMP) supplied to IP in requests for transmission.

Note that this counter does not include any datagrams counted in ipForwDatagrams.

ip_out_discards: u64

The number of output IP datagrams for which no problem was encountered to prevent their transmission to their destination, but which were discarded (e.g., for lack of buffer space).

Note that this counter would include datagrams counted in IpForwDatagrams if any such packets met this (discretionary) discard criterion.

ip_out_no_routes: u64

The number of IP datagrams discarded because no route could be found to transmit them to their destination.

Note that this counter includes any packets counted in IpForwDatagrams which meet this `no-route’ criterion.

Note that this includes any datagarms which a host cannot route because all of its default gateways are down.

ip_reasm_timeout: u64

The maximum number of seconds which received fragments are held while they are awaiting reassembly at this entity.

ip_reasm_reqds: u64

The number of IP fragments received which needed to be reassembled at this entity.

ip_reasm_oks: u64

The number of IP datagrams successfully re-assembled.

ip_reasm_fails: u64

The number of failures detected by the IP re-assembly algorithm (for whatever reason: timed out, errors, etc).

Note that this is not necessarily a count of discarded IP fragments since some algorithms (notably the algorithm in RFC 815) can lose track of the number of fragments by combining them as they are received.

ip_frag_oks: u64

The number of IP datagrams that have been successfully fragmented at this entity.

ip_frag_fails: u64

The number of IP datagrams that have been discarded because they needed to be fragmented at this entity but could not be, e.g., because their Don't Fragment flag was set.

ip_frag_creates: u64

The number of IP datagram fragments that have been generated as a result of fragmentation at this entity.

icmp_in_msgs: u64

The total number of ICMP messages which the entity received.

Note that this counter includes all those counted by icmp_in_errors.

icmp_in_errors: u64

The number of ICMP messages which the entity received but determined as having ICMP-specific errors (bad ICMP checksums, bad length, etc.

icmp_in_csum_errors: u64

This counter indicates the checksum of the ICMP packet is wrong.

Non RFC1213 field

icmp_in_dest_unreachs: u64

The number of ICMP Destination Unreachable messages received.

icmp_in_time_excds: u64

The number of ICMP Time Exceeded messages received.

icmp_in_parm_probs: u64

The number of ICMP Parameter Problem messages received.

icmp_in_src_quenchs: u64

The number of ICMP Source Quench messages received.

icmp_in_redirects: u64

The number of ICMP Redirect messages received.

icmp_in_echos: u64

The number of ICMP Echo (request) messages received.

icmp_in_echo_reps: u64

The number of ICMP Echo Reply messages received.

icmp_in_timestamps: u64

The number of ICMP Timestamp (request) messages received.

icmp_in_timestamp_reps: u64

The number of ICMP Timestamp Reply messages received.

icmp_in_addr_masks: u64

The number of ICMP Address Mask Request messages received.

icmp_in_addr_mask_reps: u64

The number of ICMP Address Mask Reply messages received.

icmp_out_msgs: u64

The total number of ICMP messages which this entity attempted to send.

Note that this counter includes all those counted by icmp_out_errors.

icmp_out_errors: u64

The number of ICMP messages which this entity did not send due to problems discovered within ICMP such as a lack of buffers. This value should not include errors discovered outside the ICMP layer such as the inability of IP to route the resultant datagram. In some implementations there may be no types of error which contribute to this counter’s value.

icmp_out_dest_unreachs: u64

The number of ICMP Destination Unreachable messages sent.

icmp_out_time_excds: u64

The number of ICMP Time Exceeded messages sent.

icmp_out_parm_probs: u64

The number of ICMP Parameter Problem messages sent.

icmp_out_src_quenchs: u64

The number of ICMP Source Quench messages sent.

icmp_out_redirects: u64

The number of ICMP Redirect messages sent. For a host, this object will always be zero, since hosts do not send redirects.

icmp_out_echos: u64

The number of ICMP Echo (request) messages sent.

icmp_out_echo_reps: u64

The number of ICMP Echo Reply messages sent.

icmp_out_timestamps: u64

The number of ICMP Timestamp (request) messages sent.

icmp_out_timestamp_reps: u64

The number of ICMP Timestamp Reply messages sent.

icmp_out_addr_masks: u64

The number of ICMP Address Mask Request messages sent.

icmp_out_addr_mask_reps: u64

The number of ICMP Address Mask Reply messages sent.

tcp_rto_algorithm: TcpRtoAlgorithmtcp_rto_min: u64

The minimum value permitted by a TCP implementation for the retransmission timeout, measured in milliseconds. More refined semantics for objects of this type depend upon the algorithm used to determine the retransmission timeout. In particular, when the timeout algorithm is rsre(3), an object of this type has the semantics of the LBOUND quantity described in RFC 793.

tcp_rto_max: u64

The maximum value permitted by a TCP implementation for the retransmission timeout, measured in milliseconds. More refined semantics for objects of this type depend upon the algorithm used to determine the retransmission timeout. In particular, when the timeout algorithm is rsre(3), an object of this type has the semantics of the UBOUND quantity described in RFC 793.

tcp_max_conn: i64

The limit on the total number of TCP connections the entity can support. In entities where the maximum number of connections is dynamic, this object should contain the value -1.

tcp_active_opens: u64

The number of times TCP connections have made a direct transition to the SYN-SENT state from the CLOSED state.

tcp_passive_opens: u64

The number of times TCP connections have made a direct transition to the SYN-RCVD state from the LISTEN state.

tcp_attempt_fails: u64

The number of times TCP connections have made a direct transition to the CLOSED state from either the SYN-SENT state or the SYN-RCVD state, plus the number of times TCP connections have made a direct transition to the LISTEN state from the SYN-RCVD state.

tcp_estab_resets: u64

The number of times TCP connections have made a direct transition to the CLOSED state from either the ESTABLISHED state or the CLOSE-WAIT state.

tcp_curr_estab: u64

The number of TCP connections for which the current state is either ESTABLISHED or CLOSE-WAIT.

tcp_in_segs: u64

The total number of segments received, including those received in error. This count includes segments received on currently established connections.

tcp_out_segs: u64

The total number of segments sent, including those on current connections but excluding those containing only retransmitted octets.

tcp_retrans_segs: u64

The total number of segments retransmitted - that is, the number of TCP segments transmitted containing one or more previously transmitted octets.

tcp_in_errs: u64

The total number of segments received in error (e.g., bad TCP checksums).

tcp_out_rsts: u64

The number of TCP segments sent containing the RST flag.

tcp_in_csum_errors: u64

[To be documented.]

Non RFC1213 field

udp_in_datagrams: u64

The total number of UDP datagrams delivered to UDP users.

udp_no_ports: u64

The total number of received UDP datagrams for which there was no application at the destination port.

udp_in_errors: u64

The number of received UDP datagrams that could not be delivered for reasons other than the lack of an application at the destination port.

udp_out_datagrams: u64

The total number of UDP datagrams sent from this entity.

udp_rcvbuf_errors: u64

[To be documented.]

Non RFC1213 field

udp_sndbuf_errors: u64

[To be documented.]

Non RFC1213 field

udp_in_csum_errors: u64

[To be documented.]

Non RFC1213 field

udp_ignored_multi: u64

[To be documented.]

Non RFC1213 field

udp_lite_in_datagrams: u64

[To be documented.]

Non RFC1213 field

udp_lite_no_ports: u64

[To be documented.]

Non RFC1213 field

udp_lite_in_errors: u64

[To be documented.]

Non RFC1213 field

udp_lite_out_datagrams: u64

[To be documented.]

Non RFC1213 field

udp_lite_rcvbuf_errors: u64

[To be documented.]

Non RFC1213 field

udp_lite_sndbuf_errors: u64

[To be documented.]

Non RFC1213 field

udp_lite_in_csum_errors: u64

[To be documented.]

Non RFC1213 field

udp_lite_ignored_multi: u64

[To be documented.]

Non RFC1213 field