procfs_core::net

Struct Snmp6

pub struct Snmp6 {

    pub ip_in_receives: u64,
    pub ip_in_hdr_errors: u64,
    pub ip_in_too_big_errors: u64,
    pub ip_in_no_routes: u64,
    pub ip_in_addr_errors: u64,
    pub ip_in_unknown_protos: u64,
    pub ip_in_truncated_pkts: u64,
    pub ip_in_discards: u64,
    pub ip_in_delivers: u64,
    pub ip_out_forw_datagrams: u64,
    pub ip_out_requests: u64,
    pub ip_out_discards: u64,
    pub ip_out_no_routes: u64,
    pub ip_reasm_timeout: u64,
    pub ip_reasm_reqds: u64,
    pub ip_reasm_oks: u64,
    pub ip_reasm_fails: u64,
    pub ip_frag_oks: u64,
    pub ip_frag_fails: u64,
    pub ip_frag_creates: u64,
    pub ip_in_mcast_pkts: u64,
    pub ip_out_mcast_pkts: u64,
    pub ip_in_octets: u64,
    pub ip_out_octets: u64,
    pub ip_in_mcast_octets: u64,
    pub ip_out_mcast_octets: u64,
    pub ip_in_bcast_octets: u64,
    pub ip_out_bcast_octets: u64,
    pub ip_in_no_ect_pkts: u64,
    pub ip_in_ect1_pkts: u64,
    pub ip_in_ect0_pkts: u64,
    pub ip_in_ce_pkts: u64,
    pub icmp_in_msgs: u64,
    pub icmp_in_errors: u64,
    pub icmp_out_msgs: u64,
    pub icmp_out_errors: u64,
    pub icmp_in_csum_errors: u64,
    pub icmp_in_dest_unreachs: u64,
    pub icmp_in_pkt_too_bigs: u64,
    pub icmp_in_time_excds: u64,
    pub icmp_in_parm_problem: u64,
    pub icmp_in_echos: u64,
    pub icmp_in_echo_replies: u64,
    pub icmp_in_group_memb_queries: u64,
    pub icmp_in_group_memb_responses: u64,
    pub icmp_in_group_memb_reductions: u64,
    pub icmp_in_router_solicits: u64,
    pub icmp_in_router_advertisements: u64,
    pub icmp_in_neighbor_solicits: u64,
    pub icmp_in_neighbor_advertisements: u64,
    pub icmp_in_redirects: u64,
    pub icmp_in_mldv2_reports: u64,
    pub icmp_out_dest_unreachs: u64,
    pub icmp_out_pkt_too_bigs: u64,
    pub icmp_out_time_excds: u64,
    pub icmp_out_parm_problems: u64,
    pub icmp_out_echos: u64,
    pub icmp_out_echo_replies: u64,
    pub icmp_out_group_memb_queries: u64,
    pub icmp_out_group_memb_responses: u64,
    pub icmp_out_group_memb_reductions: u64,
    pub icmp_out_router_solicits: u64,
    pub icmp_out_router_advertisements: u64,
    pub icmp_out_neighbor_solicits: u64,
    pub icmp_out_neighbor_advertisements: u64,
    pub icmp_out_redirects: u64,
    pub icmp_out_mldv2_reports: u64,
    pub udp_in_datagrams: u64,
    pub udp_no_ports: u64,
    pub udp_in_errors: u64,
    pub udp_out_datagrams: u64,
    pub udp_rcvbuf_errors: u64,
    pub udp_sndbuf_errors: u64,
    pub udp_in_csum_errors: u64,
    pub udp_ignored_multi: u64,
    pub udp_lite_in_datagrams: u64,
    pub udp_lite_no_ports: u64,
    pub udp_lite_in_errors: u64,
    pub udp_lite_out_datagrams: u64,
    pub udp_lite_rcvbuf_errors: u64,
    pub udp_lite_sndbuf_errors: u64,
    pub udp_lite_in_csum_errors: u64,

}

This struct holds the data needed for the IP, ICMP, TCP, and UDP management information bases for an SNMP agent.

Note that this struct is only for IPv6

For more details, see RFC1213 and SNMP counter

Fields
ip_in_receives: u64

The total number of input datagrams received from interfaces, including those received in error.

ip_in_hdr_errors: u64

The number of input datagrams discarded due to errors in their IP headers.

ip_in_too_big_errors: u64

[To be documented.]

Non RFC1213 field

ip_in_no_routes: u64

[To be documented.]

Non RFC1213 field

ip_in_addr_errors: u64

The number of input datagrams discarded because the IP address in their IP header’s destination field was not a valid address to be received at this entity.

ip_in_unknown_protos: u64

The number of locally-addressed datagrams received successfully but discarded because of an unknown or unsupported protocol.

ip_in_truncated_pkts: u64

[To be documented.]

Non RFC1213 field

ip_in_discards: u64

The number of input IP datagrams for which no problems were encountered to prevent their continued processing, but which were discarded (e.g., for lack of buffer space).

ip_in_delivers: u64

The total number of input datagrams successfully delivered to IP user-protocols (including ICMP).

Note that this counter does not include any datagrams discarded while awaiting re-assembly.

ip_out_forw_datagrams: u64

[To be documented.]

Non RFC1213 field

ip_out_requests: u64

The total number of IP datagrams which local IP user-protocols (including ICMP) supplied to IP in requests for transmission.

Note that this counter does not include any datagrams counted in ipForwDatagrams.

ip_out_discards: u64

The number of output IP datagrams for which no problem was encountered to prevent their transmission to their destination, but which were discarded (e.g., for lack of buffer space).

Note that this counter would include datagrams counted in IpForwDatagrams if any such packets met this (discretionary) discard criterion.

ip_out_no_routes: u64

The number of IP datagrams discarded because no route could be found to transmit them to their destination.

Note that this counter includes any packets counted in IpForwDatagrams which meet this `no-route’ criterion.

Note that this includes any datagarms which a host cannot route because all of its default gateways are down.

ip_reasm_timeout: u64

The maximum number of seconds which received fragments are held while they are awaiting reassembly at this entity.

ip_reasm_reqds: u64

The number of IP fragments received which needed to be reassembled at this entity.

ip_reasm_oks: u64

The number of IP datagrams successfully re-assembled.

ip_reasm_fails: u64

The number of failures detected by the IP re-assembly algorithm (for whatever reason: timed out, errors, etc).

Note that this is not necessarily a count of discarded IP fragments since some algorithms (notably the algorithm in RFC 815) can lose track of the number of fragments by combining them as they are received.

ip_frag_oks: u64

The number of IP datagrams that have been successfully fragmented at this entity.

ip_frag_fails: u64

The number of IP datagrams that have been discarded because they needed to be fragmented at this entity but could not be, e.g., because their Don't Fragment flag was set.

ip_frag_creates: u64

The number of IP datagram fragments that have been generated as a result of fragmentation at this entity.

ip_in_mcast_pkts: u64

[To be documented.]

Non RFC1213 field

ip_out_mcast_pkts: u64

[To be documented.]

Non RFC1213 field

ip_in_octets: u64

[To be documented.]

Non RFC1213 field

ip_out_octets: u64

[To be documented.]

Non RFC1213 field

ip_in_mcast_octets: u64

[To be documented.]

Non RFC1213 field

ip_out_mcast_octets: u64

[To be documented.]

Non RFC1213 field

ip_in_bcast_octets: u64

[To be documented.]

Non RFC1213 field

ip_out_bcast_octets: u64

[To be documented.]

Non RFC1213 field

ip_in_no_ect_pkts: u64

[To be documented.]

Non RFC1213 field

ip_in_ect1_pkts: u64

[To be documented.]

Non RFC1213 field

ip_in_ect0_pkts: u64

[To be documented.]

Non RFC1213 field

ip_in_ce_pkts: u64

[To be documented.]

Non RFC1213 field

icmp_in_msgs: u64

The total number of ICMP messages which the entity received.

Note that this counter includes all those counted by icmp_in_errors.

icmp_in_errors: u64

The number of ICMP messages which the entity received but determined as having ICMP-specific errors (bad ICMP checksums, bad length, etc.

icmp_out_msgs: u64

The total number of ICMP messages which this entity attempted to send.

Note that this counter includes all those counted by icmp_out_errors.

icmp_out_errors: u64

The number of ICMP messages which this entity did not send due to problems discovered within ICMP such as a lack of buffers. This value should not include errors discovered outside the ICMP layer such as the inability of IP to route the resultant datagram. In some implementations there may be no types of error which contribute to this counter’s value.

icmp_in_csum_errors: u64

This counter indicates the checksum of the ICMP packet is wrong.

Non RFC1213 field

icmp_in_dest_unreachs: u64

The number of ICMP Destination Unreachable messages received.

icmp_in_pkt_too_bigs: u64

[To be documented.]

Non RFC1213 field

icmp_in_time_excds: u64

The number of ICMP Time Exceeded messages received.

icmp_in_parm_problem: u64

The number of ICMP Parameter Problem messages received.

icmp_in_echos: u64

The number of ICMP Echo (request) messages received.

icmp_in_echo_replies: u64

The number of ICMP Echo Reply messages received.

icmp_in_group_memb_queries: u64

[To be documented.]

Non RFC1213 field

icmp_in_group_memb_responses: u64

[To be documented.]

Non RFC1213 field

icmp_in_group_memb_reductions: u64

[To be documented.]

Non RFC1213 field

icmp_in_router_solicits: u64

[To be documented.]

Non RFC1213 field

icmp_in_router_advertisements: u64

[To be documented.]

Non RFC1213 field

icmp_in_neighbor_solicits: u64

[To be documented.]

Non RFC1213 field

icmp_in_neighbor_advertisements: u64

[To be documented.]

Non RFC1213 field

icmp_in_redirects: u64

[To be documented.]

Non RFC1213 field

icmp_in_mldv2_reports: u64

[To be documented.]

Non RFC1213 field

icmp_out_dest_unreachs: u64

The number of ICMP Destination Unreachable messages sent.

icmp_out_pkt_too_bigs: u64

[To be documented.]

Non RFC1213 field

icmp_out_time_excds: u64

The number of ICMP Time Exceeded messages sent.

icmp_out_parm_problems: u64

The number of ICMP Parameter Problem messages sent.

icmp_out_echos: u64

The number of ICMP Echo (request) messages sent.

icmp_out_echo_replies: u64

The number of ICMP Echo Reply messages sent.

icmp_out_group_memb_queries: u64

[To be documented.]

Non RFC1213 field

icmp_out_group_memb_responses: u64

[To be documented.]

Non RFC1213 field

icmp_out_group_memb_reductions: u64

[To be documented.]

Non RFC1213 field

icmp_out_router_solicits: u64

[To be documented.]

Non RFC1213 field

icmp_out_router_advertisements: u64

[To be documented.]

Non RFC1213 field

icmp_out_neighbor_solicits: u64

[To be documented.]

Non RFC1213 field

icmp_out_neighbor_advertisements: u64

[To be documented.]

Non RFC1213 field

icmp_out_redirects: u64

[To be documented.]

Non RFC1213 field

icmp_out_mldv2_reports: u64

[To be documented.]

Non RFC1213 field

udp_in_datagrams: u64

The total number of UDP datagrams delivered to UDP users.

udp_no_ports: u64

The total number of received UDP datagrams for which there was no application at the destination port.

udp_in_errors: u64

The number of received UDP datagrams that could not be delivered for reasons other than the lack of an application at the destination port.

udp_out_datagrams: u64

The total number of UDP datagrams sent from this entity.

udp_rcvbuf_errors: u64

[To be documented.]

Non RFC1213 field

udp_sndbuf_errors: u64

[To be documented.]

Non RFC1213 field

udp_in_csum_errors: u64

[To be documented.]

Non RFC1213 field

udp_ignored_multi: u64

[To be documented.]

Non RFC1213 field

udp_lite_in_datagrams: u64

[To be documented.]

Non RFC1213 field

udp_lite_no_ports: u64

[To be documented.]

Non RFC1213 field

udp_lite_in_errors: u64

[To be documented.]

Non RFC1213 field

udp_lite_out_datagrams: u64

[To be documented.]

Non RFC1213 field

udp_lite_rcvbuf_errors: u64

[To be documented.]

Non RFC1213 field

udp_lite_sndbuf_errors: u64

[To be documented.]

Non RFC1213 field

udp_lite_in_csum_errors: u64

[To be documented.]

Non RFC1213 field