Module procfs_core::process

Functions and structs related to process information

The primary source of data for functions in this module is the files in a /proc/<pid>/ directory.

Structs

CoredumpFlags

See the coredump_filter() method.

Io

This struct contains I/O statistics for the process, built from /proc/<pid>/io

Limit

Limits

Process limits

MMPermissions

The permissions a process has on memory map entries.

MMapExtension

Represents the information about a specific mapping as presented in /proc/<pid>/smaps

MemoryMap

Represents an entry in a /proc/<pid>/maps or /proc/<pid>/smaps file.

MemoryMaps

Represents all entries in a /proc/<pid>/maps or /proc/<pid>/smaps file.

MemoryPageFlags

Represents the fields and flags in a page table entry for a memory page.

MountInfo

Information about a specific mount in a process’s mount namespace.

MountInfos

Information about a all mounts in a process’s mount namespace.

MountNFSStatistics

Only NFS mounts provide additional statistics in MountStat entries.

MountStat

A single entry in MountStats.

MountStats

Mount information from /proc/<pid>/mountstats.

NFSByteCounter

Represents NFS data from /proc/<pid>/mountstats under the section bytes.

NFSEventCounter

Represents NFS data from /proc/<pid>/mountstats under the section events.

NFSOperationStat

Represents NFS data from /proc/<pid>/mountstats under the section of per-op statistics.

NFSServerCaps

Namespace

Information about a namespace

Namespaces

All namespaces of a process.

Pfn

A Page Frame Number, representing a 4 kiB physical memory page

Schedstat

Provides scheduler statistics of the process, based on the /proc/<pid>/schedstat file.

SmapsRollup

Stat

Status information about the process, based on the /proc/<pid>/stat file.

StatFlags

Kernel flags for a process

StatM

Provides information about memory usage, measured in pages.

Status

Status information about the process, based on the /proc/<pid>/status file.

SwapPageFlags

Represents the fields and flags in a page table entry for a swapped page.

VmFlags

Represents the kernel flags associated with the virtual memory area. The names of these flags are just those you’ll find in the man page, but in upper case.

Enums

ClearRefs

Clearing the PG_Referenced and ACCESSED/YOUNG bits provides a method to measure approximately how much memory a process is using. One first inspects the values in the “Referenced” fields for the VMAs shown in /proc/[pid]/smaps to get an idea of the memory footprint of the process. One then clears the PG_Referenced and ACCESSED/YOUNG bits and, after some measured time interval, once again inspects the values in the “Referenced” fields to get an idea of the change in memory footprint of the process during the measured interval. If one is interested only in inspecting the selected mapping types, then the value 2 or 3 can be used instead of 1.

FDTarget

Describes a file descriptor opened by a process.

LimitValue

MMapPath

MountOptFields

Optional fields used in MountInfo

PageInfo

Represents a page table entry in /proc/<pid>/pagemap.

ProcState

Represents the state of a process.

Type Aliases

NFSPerOpStats